third_party/WebKit/LayoutTests/scrollbars/auto-scrollbar-fit-content.html

<!DOCTYPE html>
<title>Make sure a parent block that fits content expands when a child adds an auto scrollbar.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="top" style="float:left; display:none;">
  <div id="scroller" style="overflow: auto; height: 100px;">
    <div style="width: 100px; height: 200px; border: 2px solid blue;"></div>
  </div>
</div>
<script>
test(function() {
    var topDiv = document.getElementById("top");
    var scroller = document.getElementById("scroller");
    topDiv.offsetTop;
    topDiv.style.display = "block";
    assert_equals(scroller.clientWidth, scroller.scrollWidth);
});
</script>
